pyproject.toml
[project]
name = "climate-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=0.5.20",
    "cohere>=5.12.0",
    "gradio-log>=0.0.7",
    "gradio>=5.7.1",
    "humanize>=4.11.0",
    "ipython>=8.30.0",
    "langchain>=0.3.9",
    "langchain-anthropic>=0.3.0",
    "langchain-community>=0.3.8",
    "langchain-experimental>=0.3.3",
    "langchain-google-vertexai>=2.0.8",
    "langchain-nvidia-ai-endpoints>=0.3.5",
    "langchain-openai>=0.2.10",
    "langcodes>=3.5.0",
    "langgraph>=0.2.53",
    "markdown-pdf>=1.3.1",
    "msgspec>=0.18.6",
    "numpy>=1.26.4",
    "pandas>=2.2.3",
    "pathvalidate>=3.2.1",
    "pdf2docx>=0.5.8",
    "playwright>=1.49.0",
    "python-dotenv>=1.0.1",
    "redis>=5.2.0",
    "requests>=2.32.3",
    "rich>=13.9.4",
    "tiktoken>=0.8.0",
    "ua-parser>=1.0.0",
    "python-ulid>=3.0.0",
    "langchain-chroma>=0.1.4",
    "boto3>=1.35.71",
    "google-search-results>=2.4.2",
    "html2text>=2024.2.26",
    "bs4>=0.0.2",
    "ipykernel>=6.29.5",
    "unstructured>=0.16.9",
    "markdown>=3.7",
    "youtube-transcript-api>=0.6.3",
    "pip-system-certs>=4.0",
    "fastapi[standard]>=0.115.5",
    "mistune>=3.0.2",
    "pymarkdownlnt>=0.9.26",
    "stopit>=1.1.2",
    "ruff>=0.8.4",
]
